sp-44: move configs from drawer
